Added Ticks and Register methods to Level and Engine class

+++ SFML_Engine/Engine/ITickableInterface.cs
@@ -2,6 +2,6 @@
 {
     public interface ITickableInterface
     {
-        void Tick(uint deltaTime);
+        void Tick(double deltaTime);
     }
 }
